pinocchio_token

Module instructions

Modules

approve 🔒

approve_checked 🔒

burn 🔒

burn_checked 🔒

close_account 🔒

freeze_account 🔒

initialize_account 🔒

initialize_account_2 🔒

initialize_account_3 🔒

initialize_mint 🔒

initialize_mint_2 🔒

mint_to 🔒

mint_to_checked 🔒

revoke 🔒

set_authority 🔒

sync_native 🔒

thaw_account 🔒

transfer 🔒

transfer_checked 🔒

Structs

Approve

Approves a delegate.

ApproveChecked

Approves a delegate.

Burn

Burns tokens by removing them from an account.

BurnChecked

Burns tokens by removing them from an account.

CloseAccount

Close an account by transferring all its SOL to the destination account.

FreezeAccount

Freeze an Initialized account using the Mint’s freeze_authority

InitializeAccount

Initialize a new Token Account.

InitializeAccount2

Initialize a new Token Account.

InitializeAccount3

Initialize a new Token Account.

InitializeMint

Initialize a new mint.

InitializeMint2

Initialize a new mint.

MintTo

Mints new tokens to an account.

MintToChecked

Mints new tokens to an account.

Revoke

Revokes the delegate’s authority.

SetAuthority

Sets a new authority of a mint or account.

SyncNative

Given a native token account updates its amount field based on the account’s underlying lamports.

ThawAccount

Thaw a Frozen account using the Mint’s freeze_authority

Transfer

Transfer Tokens from one Token Account to another.

TransferChecked

Transfer Tokens from one Token Account to another.

Enums

AuthorityType