pinocchio_token::state::mint

Struct Mint

#[repr(C)]
pub struct Mint {
    mint_authority_flag: [u8; 4],
    mint_authority: Pubkey,
    supply: [u8; 8],
    decimals: u8,
    is_initialized: u8,
    freeze_authority_flag: [u8; 4],
    freeze_authority: Pubkey,
}

Mint data.

Fields
mint_authority_flag: [u8; 4]

Indicates whether the mint authority is present or not.

mint_authority: Pubkey

Optional authority used to mint new tokens. The mint authority may only be provided during mint creation. If no mint authority is present then the mint has a fixed supply and no further tokens may be minted.

supply: [u8; 8]

Total supply of tokens.

decimals: u8

Number of base 10 digits to the right of the decimal place.

is_initialized: u8

Is true if this structure has been initialized.

freeze_authority_flag: [u8; 4]

Indicates whether the freeze authority is present or not.

freeze_authority: Pubkey

Optional authority to freeze token accounts.

Implementations

impl Mint

pub const LEN: usize = 82usize

The length of the Mint account data.

pub fn from_account_info( account_info: &AccountInfo, ) -> Result<Ref<'_, Mint>, ProgramError>

Return a Mint from the given account info.

This method performs owner and length validation on AccountInfo, safe borrowing the account data.

pub unsafe fn from_account_info_unchecked( account_info: &AccountInfo, ) -> Result<&Self, ProgramError>

Return a Mint from the given account info.

This method performs owner and length validation on AccountInfo, but does not perform the borrow check.

Safety

The caller must ensure that it is safe to borrow the account data – e.g., there are no mutable borrows of the account data.

pub unsafe fn from_bytes(bytes: &[u8]) -> &Self

Return a Mint from the given bytes.

Safety

The caller must ensure that bytes contains a valid representation of Mint.

pub fn has_mint_authority(&self) -> bool

pub fn mint_authority(&self) -> Option<&Pubkey>

pub fn mint_authority_unchecked(&self) -> &Pubkey

Return the mint authority.

This method should be used when the caller knows that the mint will have a mint authority set since it skips the Option check.

pub fn supply(&self) -> u64

pub fn decimals(&self) -> u8

pub fn is_initialized(&self) -> bool

pub fn has_freeze_authority(&self) -> bool

pub fn freeze_authority(&self) -> Option<&Pubkey>

pub fn freeze_authority_unchecked(&self) -> &Pubkey

Return the freeze authority.

This method should be used when the caller knows that the mint will have a freeze authority set since it skips the Option check.