pinocchio_token::state::token

Struct TokenAccount

#[repr(C)]
pub struct TokenAccount {
    mint: Pubkey,
    owner: Pubkey,
    amount: [u8; 8],
    delegate_flag: [u8; 4],
    delegate: Pubkey,
    state: u8,
    is_native: [u8; 4],
    native_amount: [u8; 8],
    delegated_amount: [u8; 8],
    close_authority_flag: [u8; 4],
    close_authority: Pubkey,
}

Token account data.

Fields
mint: Pubkey

The mint associated with this account

owner: Pubkey

The owner of this account.

amount: [u8; 8]

The amount of tokens this account holds.

delegate_flag: [u8; 4]

Indicates whether the delegate is present or not.

delegate: Pubkey

If delegate is Some then delegated_amount represents the amount authorized by the delegate.

state: u8

The account’s state.

is_native: [u8; 4]

Indicates whether this account represents a native token or not.

native_amount: [u8; 8]

If is_native.is_some, this is a native token, and the value logs the rent-exempt reserve. An Account is required to be rent-exempt, so the value is used by the Processor to ensure that wrapped SOL accounts do not drop below this threshold.

delegated_amount: [u8; 8]

The amount delegated.

close_authority_flag: [u8; 4]

Indicates whether the close authority is present or not.

close_authority: Pubkey

Optional authority to close the account.

Implementations

impl TokenAccount

pub const LEN: usize = 165usize

pub fn from_account_info( account_info: &AccountInfo, ) -> Result<Ref<'_, TokenAccount>, ProgramError>

Return a TokenAccount from the given account info.

This method performs owner and length validation on AccountInfo, safe borrowing the account data.

pub unsafe fn from_account_info_unchecked( account_info: &AccountInfo, ) -> Result<&TokenAccount, ProgramError>

Return a TokenAccount from the given account info.

This method performs owner and length validation on AccountInfo, but does not perform the borrow check.

Safety

The caller must ensure that it is safe to borrow the account data – e.g., there are no mutable borrows of the account data.

pub unsafe fn from_bytes(bytes: &[u8]) -> &Self

Return a TokenAccount from the given bytes.

Safety

The caller must ensure that bytes contains a valid representation of TokenAccount.

pub fn mint(&self) -> &Pubkey

pub fn owner(&self) -> &Pubkey

pub fn amount(&self) -> u64

pub fn has_delegate(&self) -> bool

pub fn delegate(&self) -> Option<&Pubkey>

pub fn delegate_unchecked(&self) -> &Pubkey

Use this when you know the account will have a delegate and want to skip the Option check.

pub fn state(&self) -> AccountState

pub fn is_native(&self) -> bool

pub fn native_amount(&self) -> Option<u64>

pub fn native_amount_unchecked(&self) -> u64

Return the native amount.

This method should be used when the caller knows that the token is native since it skips the Option check.

pub fn delegated_amount(&self) -> u64

pub fn has_close_authority(&self) -> bool

pub fn close_authority(&self) -> Option<&Pubkey>

pub fn close_authority_unchecked(&self) -> &Pubkey

Return the close authority.

This method should be used when the caller knows that the token will have a close authority set since it skips the Option check.

pub fn is_initialized(&self) -> bool

pub fn is_frozen(&self) -> bool